+++ 3rdparty/STK/phonemes.h
@@ -1,8 +1,8 @@
 //Phonemes datas for voiceForm.dsp
 //©Romain Michon (rmichon@ccrma.stanford.edu), 2011
 //licence: STK-4.3
-STK_EXPORT float loadPhonemeGains(int index0, int index1) {
-	static float phonemeGains[32][2] =
+STK_EXPORT double loadPhonemeGains(int index0, int index1) {
+	static double phonemeGains[32][2] =
 	{{1.0, 0.0},    // eee
 		{1.0, 0.0},    // ihh
 		{1.0, 0.0},    // ehh
@@ -45,8 +45,8 @@ STK_EXPORT float loadPhonemeGains(int index0, int index1) {
   };
 	return phonemeGains[index0][index1];
 }
-STK_EXPORT float loadPhonemeParameters(int index0, int index1, int index2){
-	static float phonemeParameters[32][4][3] =
+STK_EXPORT double loadPhonemeParameters(int index0, int index1, int index2){
+	static double phonemeParameters[32][4][3] =
   {{  { 273, 0.996,  10},       // eee (beet)
       {2086, 0.945, -16}, 
       {2754, 0.979, -12},
